requirements.txt
Flask
flask-cors
tensorflow
numpy
Pillow
